Adds correct keys for EG 0104 sensor lat long
UAVBasicUniversalMetadataSet has several parsers with different keys for
'Sensor Latitude' and 'Sensor Longitude' none of which match the EG 0104
standard.  This adds the correct keys for these parameters which are the
same as those used in MISB 0601.

Rather than modify the existing parsers since legacy data may rely on
alternate keys, two new parsers are added with the suffix "Spec"
containing the correct keys: SensorLatitudeSpec and SensorLongitudeSpec.

Note that the original parsers for Sensor Latitude are using the wrong
domain for mapping; it should be +/- (2^31 - 1) according to the
specifications.